CurrencyIcon perf updates
