+++ sections/exchange/TradeCard/Cards/FooterCard.tsx
@@ -154,7 +154,6 @@ const FooterCard: React.FC = () => {
 					attemptRetry={handleApprove}
 					currencyKey={quoteCurrencyKey!}
 					currencyLabel={<NoTextTransform>{quoteCurrencyKey}</NoTextTransform>}
-					txProvider={txProvider}
 				/>
 			)}
 			{openModal === 'settle' && (
@@ -164,7 +163,6 @@ const FooterCard: React.FC = () => {
 					attemptRetry={handleSettle}
 					currencyKey={baseCurrencyKey!}
 					currencyLabel={<NoTextTransform>{baseCurrencyKey}</NoTextTransform>}
-					txProvider={txProvider}
 				/>
 			)}
 		</>
